Fix path and post-install test
We want to omit `/bin` from the install path suffix since we have
folders to "install", not just executables.

We also want to avoid running the containerd rootless setup ourselves
because that requires the appropriate version of nerdctl to be "loaded"
so that references to all the various `bin/` and `lib/` stuff are
available to the install script.